requirements.txt

flask==2.1.1
flask-migrate==3.1.0
flask-sqlalchemy==2.5.1

psycopg2-binary==2.9.3

pandas==1.4.2

python-dateutil==2.8.2

elasticsearch==8.1.2
